backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
pydantic[email]==2.9.0
python-dotenv==1.0.1
slowapi==0.1.9
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
aiosmtplib==3.0.1
jinja2==3.1.4
passlib==1.7.4
maxminddb==2.6.0
aiofiles==24.1.0
python-multipart==0.0.12
apscheduler==3.10.4